requirements.txt
datasets==2.14.3
transformers==4.31.0
pandas==2.0.3
plotly==5.15.0
umap-learn==0.5.3